Moving tests and ignore other delegates if the attribute has already been found
